+++ cattr/converters.py
@@ -16,86 +16,87 @@ V = TypeVar('V')
 
 
 @unique
-class AttrsDumpingStrategy(str, Enum):
-    """`attrs` classes dumping strategies."""
+class UnstructureStrategy(str, Enum):
+    """`attrs` classes unstructuring strategies."""
     AS_DICT = "asdict"
     AS_TUPLE = "astuple"
 
 
-DumpStratType = Union[str, AttrsDumpingStrategy]
+UnstructStratType = Union[str, UnstructureStrategy]
 
 
 class Converter:
     """Converts between structured and unstructured data."""
     def __init__(self, *, dict_factory=dict,
-                 dumping_strat: DumpStratType=AttrsDumpingStrategy.AS_DICT):
+                 unstruct_strat: UnstructStratType=
+                 UnstructureStrategy.AS_DICT):
         # Create a per-instance cache.
-        self.dumping_strat = AttrsDumpingStrategy(dumping_strat)
+        self.unstruct_strat = UnstructureStrategy(unstruct_strat)
         self._get_dis_func = lru_cache()(self._get_dis_func)
 
         # Per-instance register of to-Python converters.
-        dumps = singledispatch(self._dumps)
-        dumps.register(Enum, self._dumps_enum)
-        dumps.register(str, self._dumps_identity)
-        dumps.register(bytes, self._dumps_identity)
-        dumps.register(Sequence, self._dumps_seq)
-        dumps.register(Mapping, self._dumps_mapping)
+        unstructure = singledispatch(self._unstructure)
+        unstructure.register(Enum, self._unstructure_enum)
+        unstructure.register(str, self._unstructure_identity)
+        unstructure.register(bytes, self._unstructure_identity)
+        unstructure.register(Sequence, self._unstructure_seq)
+        unstructure.register(Mapping, self._unstructure_mapping)
 
-        self.dumps = dumps
+        self.unstructure = unstructure
 
         # Per-instance register of to-attrs converters.
         # Singledispatch dispatches based on the first argument, so we
         # store the function and switch the arguments in self.loads.
-        loads = singledispatch(self._loads_default)
-        loads.register(Any, self._loads_default)  # Bare optionals go here too.
-        loads.register(List, self._loads_list)
-        loads.register(Sequence, self._loads_list)
-        loads.register(MutableSequence, self._loads_list)
-        loads.register(MutableSet, self._loads_set)
-        loads.register(Set, self._loads_set)
-        loads.register(FrozenSet, self._loads_frozenset)
-        loads.register(Dict, self._loads_dict)
-        loads.register(Mapping, self._loads_dict)
-        loads.register(MutableMapping, self._loads_dict)
-        loads.register(Tuple, self._loads_tuple)
-        loads.register(tuple, self._loads_tuple)
-        loads.register(_Union, self._loads_union)
-        loads.register(str, self._loads_call)  # Strings are sequences.
-        loads.register(bytes, self._loads_call)  # Bytes are sequences.
-        loads.register(int, self._loads_call)
-        loads.register(float, self._loads_call)
-        loads.register(Enum, self._loads_call)
+        structure = singledispatch(self._structure_default)
+        structure.register(Any, self._structure_default)  # Bare opts here too.
+        structure.register(List, self._structure_list)
+        structure.register(Sequence, self._structure_list)
+        structure.register(MutableSequence, self._structure_list)
+        structure.register(MutableSet, self._structure_set)
+        structure.register(Set, self._structure_set)
+        structure.register(FrozenSet, self._structure_frozenset)
+        structure.register(Dict, self._structure_dict)
+        structure.register(Mapping, self._structure_dict)
+        structure.register(MutableMapping, self._structure_dict)
+        structure.register(Tuple, self._structure_tuple)
+        structure.register(_Union, self._structure_union)
+        structure.register(str, self._structure_call)  # Strings are sequences.
+        structure.register(bytes, self._structure_call)  # Bytes are sequences.
+        structure.register(int, self._structure_call)
+        structure.register(float, self._structure_call)
+        structure.register(Enum, self._structure_call)
 
-        self.loads_attrs = self.loads_attrs_fromdict
-        self._loads = loads
+        self.structure_attrs = self.structure_attrs_fromdict
+        self._structure = structure
         self._dict_factory = dict_factory
         # Unions are instances now, not classes. We use different registry.
         self._union_registry = {}
 
     @property
-    def dumping_strat(self) -> AttrsDumpingStrategy:
-        """The default way of dumping ``attrs`` classes."""
-        return (AttrsDumpingStrategy.AS_DICT
-                if self.dumps_attrs is self.dumps_attrs_asdict
-                else AttrsDumpingStrategy.AS_TUPLE)
+    def unstruct_strat(self) -> UnstructureStrategy:
+        """The default way of unstructuring ``attrs`` classes."""
+        return (UnstructureStrategy.AS_DICT
+                if self.unstructure_attrs is self.unstructure_attrs_asdict
+                else UnstructureStrategy.AS_TUPLE)
 
-    @dumping_strat.setter
-    def dumping_strat(self, val: AttrsDumpingStrategy):
-        if val is AttrsDumpingStrategy.AS_DICT:
-            self.dumps_attrs = self.dumps_attrs_asdict
+    @unstruct_strat.setter
+    def unstruct_strat(self, val: UnstructureStrategy):
+        if val is UnstructureStrategy.AS_DICT:
+            self.unstructure_attrs = self.unstructure_attrs_asdict
         else:
-            self.dumps_attrs = self.dumps_attrs_astuple
+            self.unstructure_attrs = self.unstructure_attrs_astuple
 
-    def register_dumps_hook(self, cls: Type[T], func: Callable[[T], Any]):
+    def register_unstructure_hook(self, cls: Type[T],
+                                  func: Callable[[T], Any]):
         """Register a class-to-primitive converter function for a class.
 
         The converter function should take an instance of the class and return
         its Python equivalent.
         """
-        self.dumps.register(cls, func)
+        self.unstructure.register(cls, func)
 
-    def register_loads_hook(self, cl: Type[T],
-                            func: Callable[[Any, Type], T]) -> None:
+    def register_structure_hook(self, cl: Type[T],
+                                func: Callable[[Any, Type], T]) -> None:
         """Register a primitive-to-class converter function for a type.
 
         The converter function should take two arguments:
@@ -108,60 +109,64 @@ class Converter:
         if isinstance(cl, _Union):
             self._union_registry[cl] = func
         else:
-            self._loads.register(cl, lambda t, o: func(o, t))
+            self._structure.register(cl, lambda t, o: func(o, t))
 
-    def loads(self, obj: Any, cl: Type):
+    def structure(self, obj: Any, cl: Type):
         """Convert unstructured Python data structures to structured data."""
         # Unions aren't classes, but rather instances of typing._Union now.
-        return (self._loads.dispatch(cl)(cl, obj) if not isinstance(cl, _Union)
-                else self._loads_union(cl, obj))  # For Unions.
+        return (self._structure.dispatch(cl)(cl, obj)
+                if not isinstance(cl, _Union)
+                else self._structure_union(cl, obj))  # For Unions.
 
     # Classes to Python primitives.
 
-    def _dumps(self, obj):
-        """Convert given attrs classes to their primitive equivalents."""
-        return (self.dumps_attrs(obj)
+    def _unstructure(self, obj):
+        """Convert given attrs classes to their primitive equivalents.
+
+        Other classes are passed through unchanged.
+        """
+        return (self.unstructure_attrs(obj)
                 if getattr(obj.__class__, "__attrs_attrs__", None) is not None
                 else obj)
 
-    def dumps_attrs_asdict(self, obj):
+    def unstructure_attrs_asdict(self, obj):
         """Our version of `attrs.asdict`, so we can call back to us."""
         attrs = obj.__class__.__attrs_attrs__
         rv = self._dict_factory()
         for a in attrs:
             name = a.name
             v = getattr(obj, name)
-            rv[name] = self.dumps(v)
+            rv[name] = self.unstructure(v)
         return rv
 
-    def dumps_attrs_astuple(self, obj):
+    def unstructure_attrs_astuple(self, obj):
         """Our version of `attrs.astuple`, so we can call back to us."""
         attrs = obj.__class__.__attrs_attrs__
-        return tuple(self.dumps(getattr(obj, a.name)) for a in attrs)
+        return tuple(self.unstructure(getattr(obj, a.name)) for a in attrs)
 
-    def _dumps_enum(self, obj):
+    def _unstructure_enum(self, obj):
         """Convert an enum to its value."""
         return obj.value
 
-    def _dumps_identity(self, obj):
+    def _unstructure_identity(self, obj):
         """Just pass it through."""
         return obj
 
-    def _dumps_seq(self, seq):
+    def _unstructure_seq(self, seq):
         """Convert a sequence to primitive equivalents."""
         # We can reuse the sequence class, so tuples stay tuples.
-        return seq.__class__(self.dumps(e) for e in seq)
+        return seq.__class__(self.unstructure(e) for e in seq)
 
-    def _dumps_mapping(self, mapping: Mapping):
+    def _unstructure_mapping(self, mapping: Mapping):
         """Convert a mapping of attr classes to primitive equivalents."""
         # We can reuse the mapping class, so dicts stay dicts and OrderedDicts
         # stay OrderedDicts.
-        return mapping.__class__((self.dumps(k), self.dumps(v))
+        return mapping.__class__((self.unstructure(k), self.unstructure(v))
                                  for k, v in mapping.items())
 
     # Python primitives to classes.
 
-    def _loads_default(self, cl, obj):
+    def _structure_default(self, cl, obj):
         """This is the fallthrough case. Everything is a subclass of `Any`.
 
         A special condition here handles ``attrs`` classes.
@@ -173,23 +178,24 @@ class Converter:
             return obj
         if hasattr(cl, '__attrs_attrs__'):
             # This is an attrs class
-            return self.loads_attrs(obj, cl)
+            return self.structure_attrs(obj, cl)
         # We don't know what this is, so we complain loudly.
-        msg = "Unsupported type: {0}. Register a loads hook for it.".format(cl)
+        msg = "Unsupported type: {0}. Register a structure hook for " \
+              "it.".format(cl)
         raise ValueError(msg)
 
-    def _loads_call(self, cl, obj):
+    def _structure_call(self, cl, obj):
         """Just call ``cl`` with the given ``obj``.
 
-        This is just an optimization on the ``_loads_default`` case, when we
-        know we can skip the ``if`` s. Use for ``str``, ``bytes``, ``enum``,
+        This is just an optimization on the ``_structure_default`` case, when
+        we know we can skip the ``if`` s. Use for ``str``, ``bytes``, ``enum``,
         etc.
         """
         return cl(obj)
 
     # Attrs classes.
 
-    def loads_attrs_fromtuple(self, obj: Sequence[Any], cl):
+    def structure_attrs_fromtuple(self, obj: Sequence[Any], cl):
         """Load an attrs class from a sequence (tuple)."""
         conv_obj = []  # A list of converter parameters.
         for a, value in zip(cl.__attrs_attrs__, obj):
@@ -213,13 +219,13 @@ class Converter:
                                                value)
         elif isinstance(validator, _InstanceOfValidator):
             type_ = validator.type
-            return self._loads.dispatch(type_)(type_, value)
+            return self._structure.dispatch(type_)(type_, value)
         else:
             # An unknown validator.
             return value
 
-    def loads_attrs_fromdict(self, obj: Mapping, cl):
-        """Load an attrs class from a mapping (dict)."""
+    def structure_attrs_fromdict(self, obj: Mapping, cl):
+        """Instantiate an attrs class from a mapping (dict)."""
         # For public use.
         conv_obj = obj.copy()  # Dict of converted parameters.
         for a in cl.__attrs_attrs__:
@@ -247,46 +253,47 @@ class Converter:
                                                        mapping)
         elif isinstance(val, _InstanceOfValidator):
             type_ = val.type
-            return self._loads.dispatch(type_)(type_, mapping.get(name))
+            return self._structure.dispatch(type_)(type_, mapping.get(name))
         else:
             return mapping[name]
 
-    def _loads_list(self, cl: Type[GenericMeta], obj: Iterable[T]) -> List[T]:
+    def _structure_list(self, cl: Type[GenericMeta], obj: Iterable[T])\
+            -> List[T]:
         """Convert an iterable to a potentially generic list."""
         if not cl.__args__ or cl.__args__[0] is Any:
             return [e for e in obj]
         else:
             elem_type = cl.__args__[0]
-            conv = (self._loads.dispatch(elem_type)
+            conv = (self._structure.dispatch(elem_type)
                     if not isinstance(elem_type, _Union)
-                    else self._loads_union)
+                    else self._structure_union)
             return [conv(elem_type, e) for e in obj]
 
-    def _loads_set(self, cl: Type[GenericMeta], obj: Iterable[T])\
+    def _structure_set(self, cl: Type[GenericMeta], obj: Iterable[T])\
             -> MutableSet[T]:
         """Convert an iterable into a potentially generic set."""
         if not cl.__args__ or cl.__args__[0] is Any:
             return set(obj)
         else:
             elem_type = cl.__args__[0]
-            conv = (self._loads.dispatch(elem_type)
+            conv = (self._structure.dispatch(elem_type)
                     if not isinstance(elem_type, _Union)
-                    else self._loads_union)
+                    else self._structure_union)
             return {conv(elem_type, e) for e in obj}
 
-    def _loads_frozenset(self, cl: Type[GenericMeta], obj: Iterable[T])\
+    def _structure_frozenset(self, cl: Type[GenericMeta], obj: Iterable[T])\
             -> FrozenSet[T]:
         """Convert an iterable into a potentially generic frozenset."""
         if not cl.__args__ or cl.__args__[0] is Any:
             return frozenset(obj)
         else:
             elem_type = cl.__args__[0]
-            conv = (self._loads.dispatch(elem_type)
+            conv = (self._structure.dispatch(elem_type)
                     if not isinstance(elem_type, _Union)
-                    else self._loads_union)
+                    else self._structure_union)
             return frozenset([conv(elem_type, e) for e in obj])
 
-    def _loads_dict(self, cl: Type[GenericMeta], obj: Mapping[T, V])\
+    def _structure_dict(self, cl: Type[GenericMeta], obj: Mapping[T, V])\
             -> Dict[T, V]:
         """Convert a mapping into a potentially generic dict."""
         if not cl.__args__ or cl.__args__ == (Any, Any):
@@ -294,32 +301,32 @@ class Converter:
         else:
             key_type, val_type = cl.__args__
             if key_type is Any:
-                val_conv = (self._loads.dispatch(val_type)
+                val_conv = (self._structure.dispatch(val_type)
                             if not isinstance(val_type, _Union)
                             else self._loads_union)
                 return {k: val_conv(val_type, v) for k, v in obj.items()}
             elif val_type is Any:
-                key_conv = (self._loads.dispatch(key_type)
+                key_conv = (self._structure.dispatch(key_type)
                             if not isinstance(key_type, _Union)
                             else self._loads_union)
                 return {key_conv(key_type, k): v for k, v in obj.items()}
             else:
-                key_conv = (self._loads.dispatch(key_type)
+                key_conv = (self._structure.dispatch(key_type)
                             if not isinstance(key_type, _Union)
-                            else self._loads_union)
-                val_conv = (self._loads.dispatch(val_type)
+                            else self._structure_union)
+                val_conv = (self._structure.dispatch(val_type)
                             if not isinstance(val_type, _Union)
                             else self._loads_union)
                 return {key_conv(key_type, k): val_conv(val_type, v)
                         for k, v in obj.items()}
 
-    def _loads_union(self, union: _Union, obj: Any):
-        """Deal with converting a union.
+    def _structure_union(self, union: _Union, obj: Any):
+        """Deal with converting a union."""
 
-        Note that optionals are unions that contain NoneType. We check for
-        NoneType early and handle the case of obj being None, so
-        disambiguation functions don't need to handle NoneType.
-        """
+        # Note that optionals are unions that contain NoneType. We check for
+        # NoneType early and handle the case of obj being None, so
+        # disambiguation functions don't need to handle NoneType.
+
         # Check the union registry first.
         handler = self._union_registry.get(union)
         if handler is not None:
@@ -335,15 +342,15 @@ class Converter:
                 other = (union_params[0] if union_params[1] is NoneType
                          else union_params[1])
                 # We can't actually have a Union of a Union, so this is safe.
-                return self._loads.dispatch(other)(other, obj)
+                return self._structure.dispatch(other)(other, obj)
 
         # Getting here means either this is not an optional, or it's an
         # optional with more than one parameter.
         # Let's support only unions of attr classes for now.
         cl = self._get_dis_func(union)(obj)
-        return self._loads.dispatch(cl)(cl, obj)
+        return self._structure.dispatch(cl)(cl, obj)
 
-    def _loads_tuple(self, tup: Type[Tuple], obj: Iterable):
+    def _structure_tuple(self, tup: Type[Tuple], obj: Iterable):
         """Deal with converting to a tuple."""
         tup_params = tup.__args__
         has_ellipsis = (tup_params and tup_params[-1] is Ellipsis)
@@ -353,15 +360,15 @@ class Converter:
         if has_ellipsis:
             # We're dealing with a homogenous tuple, Tuple[int, ...]
             tup_type = tup_params[0]
-            conv = (self._loads.dispatch(tup_type)
+            conv = (self._structure.dispatch(tup_type)
                     if not isinstance(tup_type, _Union)
-                    else self._loads_union)
+                    else self._structure_union)
             return tuple(conv(tup_type, e) for e in obj)
         else:
             # We're dealing with a heterogenous tuple.
-            return tuple(self._loads.dispatch(t)(t, e)
+            return tuple(self._structure.dispatch(t)(t, e)
                          if not isinstance(t, _Union)
-                         else self._loads_union(t, e)
+                         else self._structure_union(t, e)
                          for t, e in zip(tup_params, obj))
 
     def _get_dis_func(self, union: Type) -> Callable[..., Type]: